refactor(server): refactor http+ws server
