+++ src/server/app.js
@@ -1,66 +1,11 @@
-import koa from 'koa';
+import app from './http';
+import handler from './ws';
+
 import http from 'http';
 import ws from 'ws';
-import { get } from 'koa-route';
-import mount from 'koa-mount';
-import assets from './routes/assets';
-
-const app = koa();
-
-app
-  .use(mount(assets))
-  .use(get('/', function* () {
-    this.body = '<link rel=stylesheet href=style.css /><div id=mount></div><canvas id=canvas width=500 height=500></canvas><script src=bundle.js></script>';
-  }));
 
 const server = http.createServer(app.callback());
-const wsServer = new ws.Server({server});
+const wss = new ws.Server({server});
+wss.on('connection', handler);
 
 server.listen(3000);
-
-let uid = 0;
-const clients = new Set();
-
-function broadcast(action) {
-  const message = JSON.stringify(action);
-  [...clients].forEach(client => {
-    client.send(message);
-  });
-}
-
-function act(type, payload, meta) {
-  return {
-    type,
-    payload,
-    meta,
-  };
-}
-
-wsServer.on('connection', client => {
-  client.id = uid++;
-  broadcast(act('ADD_USER', {
-    id: client.id,
-  }));
-  clients.add(client);
-  client.on('message', message => {
-    const action = JSON.parse(message);
-    switch (action.type) {
-      case 'ADD_MESSAGE':
-      setTimeout(() => {
-        broadcast(act('ADD_MESSAGE', {
-          message: action.payload.message,
-          author: client.id,
-          date: new Date(),
-        }, {
-          done: true,
-          id: action.meta.id,
-        }));
-      }, 1000);
-    }
-  }).on('close', () => {
-    clients.delete(client);
-    broadcast(act('DEL_USER', {
-      id: client.id,
-    }));
-  });
-});
